+++ engine/private/core/engine.cpp
@@ -1,0 +1,7 @@
+#include "core/engine.h"
+
+// 💀
+
+Core core;
+
+ma_engine engine;
