+++ engine/public/core/ember_core.h
@@ -2,10 +2,9 @@
 
 #include "core/component/camera.h"
 #include "core/component/component.h"
-#include "core/file_system.h"
 #include "core/renderer/mesh.h"
 #include "core/renderer/shader.h"
-#include "core/time_manager.h"
+#include "core/audio/ember_audio.h"
 
 
 /*!
@@ -56,30 +55,7 @@ private:
 
     @version 0.0.1
 */
-struct Core {
 
-    struct {
-        bool bDevelopmentLogging = true;
-    } Metrics;
-
-    struct {
-        int width;
-        int height;
-        const char* title           = "[EMBER_ENGINE] - Window";
-        const SDL_DisplayMode* data = nullptr;
-
-    } Window;
-
-    struct {
-        SDL_AudioDeviceID device_id = 0;
-        SDL_AudioSpec spec;
-        float global_volume = 5.0f;
-        Ember_VFS ember_vfs;
-    } Audio;
-
-
-    TimeManager* Time = nullptr;
-};
 
 
 /*!
@@ -146,105 +122,6 @@ void CloseWindow();
 */
 bool InitWindow(const char* title, int width, int height, RendererType type, Uint64 flags = 0);
 
-
-/*!
-
-    @brief Initialize Sound Engine
-    - mini-audio backend
-
-    @version 0.0.3
-    @return bool
-*/
-bool InitAudio();
-
-/*!
-
-    @brief Close Sound Engine and free allocated resources
-
-
-    @version 0.0.3
-    @return void
-*/
-void CloseAudio();
-
-
-/*!
-
-    @brief Load Audio and keep track of it
-
-    @param file_Path Path to the audio file
-
-    @version 0.0.3
-    @return Audio allocated audio memory struct
-*/
-Audio* Mix_LoadAudio(const std::string& file_Path);
-
-/*!
-
-    @brief Play audio
-
-    @param music Valid Audio struct pointer
-    @param loop Play audio in `loop` or not
-
-    @version 0.0.3
-    @return void
-*/
-void Mix_PlayAudio(Audio* audio, bool loop = false);
-
-
-/*!
-
-    @brief Pause audio and fade out
-
-    @param audio Valid Audio struct pointer
-
-    @version 0.0.3
-    @return void
-*/
-void Mix_PauseAudio(Audio* audio);
-
-/*!
-
-    @brief Change the audio volume
-
-    @param audio Valid Audio struct pointer
-    @param volume Audio volume from 0.0 to 1.0
-
-    @version 0.0.3
-    @return void
-*/
-void Mix_SetVolume(Audio* audio, float volume);
-
-/*!
-
-    @brief Change the [audio engine] global volume
-
-    @param volume Audio volume from 0 to 100
-
-    @version 0.0.3
-    @return void
-*/
-void Mix_SetGlobalVolume(float volume);
-
-/*!
-
-    @brief Unload Audio allocated memory
-    - Manually clear allocated audio or at the end will be freed automatically
-
-    @see CloseAudio
-
-    @param audio Valid Audio struct pointer
-
-    @version 0.0.3
-    @return void
-*/
-void Mix_UnloadAudio(Audio* audio);
-
-// TODO: create a Resource Manager
-inline std::unordered_map<std::string, Audio*> audios;
-
 static Renderer* renderer = nullptr;
 
-extern ma_engine engine;
 
-extern Core core;
